+++ database/migrations/2023_11_26_233145_create_daily_task_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('daily_task', function (Blueprint $table) {
             $table->id();
             $table->string('link');
-            $table->integer('price');
+            $table->float('price', 10, 3);
             $table->string('description');
             $table->string('plan');
             $table->string('level');
